refactor: 인기테마 조회 메소드 객체 변경
